/tickets API: fix gear order dates
The CashNet requests no longer appear to include the
'created_at' or 'updated_at' fields. Thus when we tried to
parse those (undefined) values into dates, we got back dates
corresponding with the Unix epoch (1/1/1970). This lead to
all entries on the 'gear orders' spreadsheet - and indeed the
corresponding values in the database - being inaccurate.

Because this field no longer exists, we simply take the
current server time to be the default value for these fields.