Team Comp: Fix mongo edge case error.
There's a mongo edge case that will cause an error if the 'myTeam'
service is called when the current user is not on a team. It appears
to be due to Mongo and Meteor disagreeing on how to handle the query
when `user.teamId` is undefined. Thus we only run the query if the
user is actually on a team.

For the time being, additional print statements are being added to
help provide insight if this continues to be a problem, but they
could/should be removed at a later date or at least turned into
'debug' level logs.