meteor: update to meteor@1.8.3
REQUIRED ACTION FOR DEVELOPERS:
This update moves to a new version of mongodb which is not backwards
compatible. You will likely have to run "meteor reset" to reset the
database before starting the server for the first time. If you have
data you would like to retain, you may be able to make a backup
with mongodump and restore it with mongorestore.