Migrations: add 'inPersonAllowed' migration
For the time being, update all users such that inPersonAllowed=false.
We will have to manually fix this setting for some people, but it is
easier and safer to do that for the few eligible users we currently
have than to risk adding buggy logic on short notice to properly set
this value for all users.