GameMode can only be changed when registration is open
We need an accurate headcount a few days ahead of the hunt for
ordering food and the like. Thus the gameMode (in-person vs
virtual) must be set in stone for players a few days before the
game. The ability to toggle this setting is now tied to the
registration closed date.