Split registration toggle into in-person and virtual
We need to be able to turn off in-person registration ahead of the
game, either when capacity is reached or the deadline is up. However,
we'd still like to keep virtual registration open until closer to the
event itself.

This commit splits the functionality such that both options can be
enabled or disabled independently. Turning off in-person registration
also disables the user's ability to toggele their profile between
in-person and virtual. Team toggles are unaffected because the team
setting has more to do with puzzle delivery through the system than it
does with the players' physical locations.