Add warning labels to input fields in Puzzle entry
Add warning labels to catch some basic mistakes in the puzzle entry
fields on the admin page (eg empty fields, non-numeric values in
numeric inputs, etc.). This is a basic first line of defense in case
an admin fails to enter all puzzle information in. Note that it is NOT
blocking, so a puzzle with empty fields is still allowed. These are
just visual nudges.