Handle 'user.roles' being undefined
Occasionally, a User object does not have a .roles array when a page
loads. This causes the `hasRole` method to fail and the page will not
load. As a workaround, we can return false from any `hasRole` call
where the roles property doesn't exist. However, it is not yet well
understood why the object is in a partial state at this point.

There seem to be two places this fails: Authed.jsx and
TopBar.jsx. When the user object is not fully loaded, both of these
fail on page load. It seems the User object in the browser's database
is incomplete, potentially.