Volunteer: don't show full CompletedPuzzle comp
There's no real reason for a volunteer to need to see the details
included with the CompletedPuzzle after starting a team's timer. It's
also very unlikely that anyone stays on this page long enough to care
or notice the difference.

Removing this dependency will allow for some small cleanup and
improvements to the Player-facing component.