Hide puzzle answer if team timed out
Per player request, if a puzzle times out, don't immediately display
the correct answer. (Some teams reported being nearly done when the
timer expired and then having the answer "spoiled" for them.) They
will need the answer for the meta puzzle, but not necessarily
immediately.

This change affects only timed-out completed puzzles: when loaded, if
a puzzle is timed out, display a "Reveal" button instead of the
answer. Clicking the button shows the answer. The answer cannot be
hidden again unless the component is reset (e.g. via a page reload).