Leaderboard: fix isAdmin check and missing import
The leaderboard was relying on an isAdmin check that had been
converted to an async function during the Meteor@3 migration. On the
client side, this check is now done inside the component in a
synchronous manner and the result is passed down into child
components. On the server, an 'await' has been added.

Additionally, a missing import was fixed.