Add inactivity timeout to Volunteer page
Traditional Volunteer workflow involves scanning a QR code which opens
a new tab, allowing them to start the puzzle. This leads to multiple
open tabs in the background after a short amount of time.

While we can't close a tab that we didn't `window.open()`, we can at
least navigate to a page that does not have an active websocket
connection to the Meteor backend. This doesn't save any time for our
volunteers, but it will hopefully reduce server load as more and more
teams start puzzles.

This is implemented as a two-step timer: after the first timer
expires, a warning message is displayed. If the user takes no action,
then the page will be redirected to the "inactive" HTML page after an
additional timeout (similar to the "you're about to be logged out"
messages on a banking website).