Move all leader profile photos to AWS
We hit the max "slug" size for Heroku with the addition of the three
newest leaders. While storing images in the repo may have been okay
early on when there were few of them, we've apparently reached
capacity and should move these to a proper CDN. This is a
long-standing but low-prority task I've had that seems necessitated by
the latest changes.

All photos have been moved into S3 (and Millie's photo is now in the
same directory as the others).